pyproject.toml
[project]
name = "ny-taxi-warehouse"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dbt-bigquery>=1.11.0",
    "dbt-core>=1.11.5",
    "google-cloud-bigquery>=3.40.1",
    "google-cloud-storage>=3.1.1",
    "httpx>=0.28.1",
    "pandas>=3.0.0",
    "pyarrow>=23.0.1",
    "python-dotenv>=1.2.1",
]
